requirements.txt
pillow>=9.0.0
numpy>=1.20.0
opencv-python>=4.5.0
gdown>=4.4.0

# Les packages ci-dessous sont commentés car ils causent des problèmes d'installation
# torch>=1.8.0
# realesrgan>=0.3.0
# basicsr>=1.4.2
# facexlib>=0.2.5

# Pour la suppression d'arrière-plan (fonctionne bien)
rembg
onnxruntime